+++ crates/quickjs-wasm-rs/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [2.0.1] - 2023-09-08
+## [2.0.1] - 2023-09-11
 
 ### Fixed